Refactor Pipeline REST API endpoint
Move pipeline submission endpoint from
api/scheduler to api/pipeline/schedule
