+++ elyra/pipeline/__init__.py
@@ -14,7 +14,7 @@
 # limitations under the License.
 #
 
-from .handlers import PipelineExportHandler
+from .handlers import PipelineExportHandler, PipelineSchedulerHandler
 from .parser import PipelineParser
 from .pipeline import Operation, Pipeline
 from .processor import PipelineProcessorRegistry, PipelineProcessorManager, PipelineProcessor, PipelineProcessorResponse
