+++ elyra/pipeline/handlers.py
@@ -18,17 +18,22 @@ import json
 from notebook.base.handlers import APIHandler
 from .parser import PipelineParser
 from .processor import PipelineProcessorManager
+from tornado import web, gen
 from ..util.http import HttpErrorMixin
 
 
 class PipelineExportHandler(HttpErrorMixin, APIHandler):
     """Handler to expose REST API to export pipelines"""
 
+    @web.authenticated
+    @gen.coroutine
     def get(self):
         msg_json = dict(title="Operation not supported.")
         self.write(msg_json)
         self.flush()
 
+    @web.authenticated
+    @gen.coroutine
     def post(self, *args, **kwargs):
         self.log.debug("Pipeline Export handler now executing post request")
 
@@ -51,8 +56,32 @@ class PipelineExportHandler(HttpErrorMixin, APIHandler):
         self.write(json_msg)
         self.flush()
 
-    def __artifact_list_to_str(self, pipeline_array):
-        if not pipeline_array:
-            return "None"
-        else:
-            return ','.join(pipeline_array)
+
+class PipelineSchedulerHandler(HttpErrorMixin, APIHandler):
+    """Handler to expose method calls to execute pipelines as batch jobs"""
+
+    @web.authenticated
+    @gen.coroutine
+    def get(self):
+        msg_json = dict(title="Operation not supported.")
+        self.write(msg_json)
+        self.flush()
+
+    @web.authenticated
+    @gen.coroutine
+    def post(self, *args, **kwargs):
+        self.log.debug("Pipeline SchedulerHandler now executing post request")
+
+        pipeline_definition = self.get_json_body()
+
+        self.log.debug("JSON payload: %s", pipeline_definition)
+
+        pipeline = PipelineParser.parse(pipeline_definition)
+
+        response = PipelineProcessorManager.process(pipeline)
+        json_msg = json.dumps(response.to_json())
+
+        self.set_status(200)
+        self.set_header("Content-Type", 'application/json')
+        self.finish(json_msg)
+        self.flush()
